requirements.txt
aiohttp
discord.py
discord.py[voice]
gtts
langdetect
python-dotenv
pyttsx3
PyYAML
Requests
nest_asyncio
g4f